allow overwritte interface method
Summary:
a.k.a: so close to have "pystemd-run" implemented! (look at the example file)

Sometimes (like really not that common) a method have weird complex
types, (the variant type). now with this we can just writte our own method with
python.

The added bonus is that now we can simply start a transient unit (this is good
for now, but in the future we may want to add more fields to the unit).

Reviewed By: davide125

Differential Revision: D6258129

fbshipit-source-id: 64f074752ee098746903ff18533465c6860b95fd