add legal info necesary for deploy wheel and source dist
Summary:
This diff is meant to enable us (from the legal point of view) distribute pystemd in a wheel and a source dist.

Wheel files has everything compiled inside of them and are basically code. they do provide a METADATA section that allow for only 1 license file, for that, we added the LICENSE.txt that specify the license and also direct the user on where to get them (the main repo)

source distribution does not ship with compiled code, that menas that users needs to build their own code, but allow us to include ANY data we want in the main directory, for those effects i created the MANIFEST.in that allow to include the LICENSE files.

Reviewed By: postal007

Differential Revision: D6491105

fbshipit-source-id: 9b6663bd6275c6ebe5f5b4e9492c91cbb1cd5bc0