create pystemd.run the cheap clone of systemd-run for python
Summary:
they say imitation is the biggest form of praise , i don't really know about that, but this is suppose to be cool.

so basically this is systemd-run in the shape of pystemd.run, this allow us in s simple form execute transient units, this is really like subprocess.call but with steroids (granted that it does not do everything that  subprocess.call can right now. but its a awesome start).

so lets get started on the list of features:

now you can do:

```
    >>> import pystemd
    >>> pystemd.run([b'/bin/sleep', b'42'])
```
and start a unit that executes /bin/sleep 42. the control will be returned to you with no output... so unix philosophy that hirts

also you can do:

2.- executes `/bin/sleep 42` but returns the units

    >>> import pystemd
    >>> unit = pystemd.run([b'/bin/sleep', b'42'], remain_after_exit=True)
    >>> unit
    <pystemd.systemd1.unit.Unit at 0x7f8c460695c0>
    >>> unit.Service.MainPID
    66244
    >>> # ... waiting 42 seconds later
    >>> unit.Service.MainPID
    0

it will execute the same command but returns a unit object that you can query for live changes.

3.- executing `/bin/sleep 42` as other user and waiting for process to end

    >>> import pystemd
    >>> unit = pystemd.run([b'/bin/sleep', b'42'], remain_after_exit=True,
        wait=True, user=b'aleivag'
    ) # ... waiting 42 seconds later
    >>> unit.Service.MainPID
    0

now the method will not give you control until you finish

4.- want all of this, but also ruining inside a different machine, no problem

    >>> import pystemd
    >>> pystemd.run(
        [b'/bin/env'],
        machine=b'miniarch',
        remain_after_exit=True,
        wait=True,
        user_mode=True,
        nice=3,
        env={b'SUPER_SECRET_PASSW': b'1234'}
        )
    <pystemd.systemd1.unit.Unit at 0x7f7d5894b668>

goodies:

  # You can execute a command as another user.
  # Ability of all pystemd to talk to systemd without accessing the bus
  # You can execute command on the context of a machine, or user.
  # You can execute a command, and that command will be executed by systemd, so if your script dies, it can resume from where it left of, if it knows the name of the unit.
  # you can now call methods that returns more than one answer and get them all!

Reviewed By: zfjagann

Differential Revision: D6310411

fbshipit-source-id: 41238c074412a19af24334c76b0931bc669b87e7