cleanup some memory leaks
Summary:
I have identified some parts of the pystemd where we don't deallocate pointers (like char*), expecting that cython would do that for us when we return... it turns out that cython can't actually do it, because that is now how c works. so we will do it on our own...

this is actually a good way to start cleaning memory leaks. thankfully our codebase is small and most of the work is done by libsystemd itself, so it should be fairly easy to fix any memory problem we have in the future.

Reviewed By: davide125

Differential Revision: D8274686

fbshipit-source-id: ae1e337faf4b22fe67e1f5aca889f87dab12eda3