- Enable booted image build in lego castle
Summary:
Since *castle now uses initcastle to isolate the processes for a given job, we had to figure out some kind of workaround for buildinator since it wasn't able to contact the parent hosts' systemd in order to register the container.

This diff adds the options to not register the container on boot, so it would run as a completely isolated container. However, in doing this, it also limits how we can communicate into the container.

My first attempt at this was to create a local-to-buildinator tmpfs and mount that inside the container at /run so the sockets created by systemd inside the container would be exposed to the outside world, but this caused issues with verifying the validity of the calling process inside the container.

After working on this for a while with mokomull, we found there was actually a bug in pystemd (D16743979) with how the library was setting up the new sd_bus object as a client. We found that it was actually a cleaner solution to use the x-machine-unix address type, as this forks a new process into the namespace of the container, which gets around the valid pid issue.

Tested this locally on dev server, and created an ephemeral buildinator package to ensure it works in legocastle/initcastle environment, and this should be good to go.

Once Matt lands his diff regarding pystemd, I can rebase this and get rid of those changes out of this diff, but keeping in for now for completeness sake.

Reviewed By: zeroxoneb

Differential Revision: D16753675

fbshipit-source-id: c0d904ffee360d55446a813a8efc5ed81aaf1616