Run black on tests with the latest version (#40)
Summary:
In Travis, we are using black 19.10b0, while the one I was running in my dev machine is 19.3b0. Let's blacken the file with the latest version of black, so our OS CI is green!

We might want to lock black to the same version we run internally
Pull Request resolved: https://github.com/facebookincubator/pystemd/pull/40

Differential Revision: D19623956

Pulled By: aleivag

fbshipit-source-id: 2feebc1f4a85a5b8172dfab19ab712970f184cb3