add a LIBSYSTEMD_VERSION macro for gating new functions (#58)
Summary:
This allows gating of function-symbols based on the builder's own libsystemd version.

Unfortunately cython doesn't translate both branches of the IF-ELSE from *.pyx into the intermediate *.c files, so if you build source zips on a host with a newer libsystemd, that version's symbols will be expected to be available for the final compile.

I guess that users needing to support older systemd installations are most likely going to be starting from a clean .git checkout and using cython, not a dist .zip, so perhaps that's not a big deal?

I didn't make a note of it in the README since we only seem to have one sd_ function that's not totally in the field yet and eventually these older versions should die off...

Fixes https://github.com/facebookincubator/pystemd/issues/26

Pull Request resolved: https://github.com/facebookincubator/pystemd/pull/58

Reviewed By: davide125

Differential Revision: D25989028

Pulled By: aleivag

fbshipit-source-id: c7d69b20691f7dfc1dd0f7e5c42e02bf0a117b73